Creating a ContainerTestBase to isolate tests common to IContainer implementation classes.
While creating this class, I found that we already have a ContainerTestBase (in the reflex.display package).  I initially tried to add my test cases to that class.  However, that class also extends StatefulTestBase.  Since not all implementations of IContainer have a "states" property, trying to use that version of ContainerTestBase was causing problems.  We could probably go further and claim that this is likely an improper extension (assuming a container has states is, at least currently, incorrect).  For now, I renamed the version in reflex.display to "StatefulContainerTestBase" to clarify that it combines the two concepts.  Going forward, we may want to move toward a reuse-by-composition arrangement, rather than a reuse-by-inheritance arrangement.