Creating SelectionTest class to test Selection class.  These tests discovered an issue in the _not_ changed scenario: if the current _selectedItem is "abc" ... and you then try to set the selectedItem to "abc" (i.e., same value, no true change), a selectedItemChange event was being dispatched.  I added code at the beginning of the "set selectedItem(...)" method to "return" if the supplied value is the same as the current value.
One curious thing about this class is that, while there is a selectedItems (plural) property, it does not appear that the IList could ever contain more than a single item.  The only "public" way to add an item to the IList is to call "set selectedItem(...)" (singular).  But, the first thing "set selectedItem(...)" does is to call removeAll() on the IList, prior to  adding this new item to the IList.