Adding tests to BitmapDisplayMeasurableTest class.  While testing the "set style(value:*) method in BitmapDisplay, I noticed a couple of things.
First, there is an assumption being made that any values passed-in will adhere to the delimiters ":" and ";".  However, there is nothing to insure or inform about whether these assumptions were met.  If a developer calls this method, they likely expect that, absent an error, their call succeeded.  However, if they used the wrong delimiters, the method would fail silently.  I added code (including a new package-level "check..." method) to throw an error if the delimiter assumptions are not met.

Second, while the method signature allows any parameter type to be passed in (*), it actually only does something if the parameter is of type String.  I'm guessing that the desire here is to add handling for other types in the future (?).  In the meantime, however, it would probably be helpful to inform when an unsupported parameter type is supplied (similar thinking as before - don't let the developer think the call succeeded when it didn't).

There are obviously other ways these concerns could be handled.  As one example, perhaps "set style(...)" could take a PropertyValue object, rather than a delimited string.  So, please think about whether you'd prefer to handle these items in some other manner when considering this pull request.